packages/vectorstore/pyproject.toml
[project]
name = "astra-vectorstore"
version = "0.1.0"
description = "Astra FAISS wrapper + sentence-transformers embeddings."
requires-python = ">=3.11,<3.13"
dependencies = [
    "faiss-cpu>=1.8",
    "numpy>=1.26",
    "sentence-transformers>=3.0",
    "pydantic>=2.7,<3.0",
    "astra-schemas",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["astra_vectorstore"]
